+++ client/projects/client/src/app/modules/main/components/layout/sidebar/sidebar.component.ts
@@ -1,9 +1,11 @@
 import { Component, OnInit } from "@angular/core";
 import { version } from '../../../../../../../../../package.json';
-import { map, Observable, of, switchMap } from "rxjs";
+import { from, map, mergeMap, Observable, of } from "rxjs";
 import { DiscoItem, DiscoItems } from "stanza/protocol";
 import { XmppService } from "../../../../../lib/services/xmpp.service";
-// import { XmppService } from "../../../services/xmpp.service";
+import { Store } from "@ngrx/store";
+import { AppState } from "../../../../../store/store";
+import { ChatRoomAdd } from "../../../../../store/chat/actions";
 
 @Component({
     selector: 'app-sidebar',
@@ -13,11 +15,32 @@ export class SidebarComponent implements OnInit {
     version: string;
     rooms$: Observable<DiscoItem[]> = of([]);
 
-    constructor(private xmpp: XmppService) { }
+    constructor(private xmpp: XmppService, private store: Store<AppState>) { }
     ngOnInit(): void {
         this.version = version
-        this.rooms$ = this.xmpp.queryRoomsOnline().pipe(
-            map((data) => (data.disco as DiscoItems).items || [])
+        this.rooms$ = this.loadRoomsAndDispatch();
+    }
+
+    private loadRoomsAndDispatch(): Observable<DiscoItem[]> {
+        return this.xmpp.queryRoomsOnline().pipe(
+            map((data) => (data.disco as DiscoItems).items || []),
+            mergeMap((rooms: DiscoItem[]) => from(rooms)),
+            mergeMap((room: DiscoItem) =>
+                this.xmpp.getRoomInfo(room.jid as string).pipe(
+                    map((info) => ({ room, info }))
+                )
+            ),
+            map(({ room, info }) => {
+                this.dispatchRoomInfo(room.jid as string, info);
+                return room;
+            }),
+            map((room) => [room])
         );
+    }
+
+    private dispatchRoomInfo(jid: string, info: any): void {
+        this.store.dispatch(ChatRoomAdd({
+            payload: { id: jid, info }
+        }));
     }
 }