+++ client/projects/client/src/app/store/chat/reducers/rooms.ts
@@ -1,7 +1,7 @@
-import { createReducer, createSelector, on } from "@ngrx/store";
+import { createReducer, on } from "@ngrx/store";
 import { EntityAdapter, EntityState, createEntityAdapter } from "@ngrx/entity";
-import * as actions from './actions';
-import { Room } from "./def";
+import * as actions from '../actions';
+import { Room } from "../def";
 
 export interface RoomsState extends EntityState<Room> { }
 export const RoomsAdapter: EntityAdapter<Room> = createEntityAdapter<Room>();